client-api: Rename mutual_rooms module to get_mutual_rooms
Because our API endpoint modules always use a verb to reflect that they
perform an action.
