+++ crates/ruma-client-api/src/membership.rs
@@ -3,6 +3,7 @@
 pub mod ban_user;
 pub mod forget_room;
 pub mod get_member_events;
+pub mod get_mutual_rooms;
 pub mod invite_user;
 pub mod join_room_by_id;
 pub mod join_room_by_id_or_alias;
@@ -10,7 +11,6 @@ pub mod joined_members;
 pub mod joined_rooms;
 pub mod kick_user;
 pub mod leave_room;
-pub mod mutual_rooms;
 pub mod unban_user;
 
 use ruma_common::{OwnedUserId, ServerSignatures, thirdparty::Medium};
