+++ crates/ruma-client-api/src/membership/get_mutual_rooms.rs
